add attachments in expenses

+++ src/routes/expense/+page.server.ts
@@ -5,6 +5,7 @@ import { itemSchema } from '$lib/validations.js'; // Assuming itemSchema is a ge
 import { getPaginationData } from '$lib/utils.js';
 import { initForm, sendMessage, validateAction } from '$lib/superforms';
 import { title } from './upsert/utils.js';
+import { delFile } from '$lib/server/filesystem.js';
 
 export async function load(event) {
   const form = await initForm(itemSchema); // Generic form for deletion
@@ -32,6 +33,20 @@ export const actions = {
   async default(event) {
     const form = await validateAction(event, itemSchema);
     if (!form.valid) return form.error;
+
+    const expense = await db.query.Expenses.findFirst({
+      where: eq(Expenses.id, form.data.id),
+      columns: { attachmentUrls: true },
+    });
+    if (!expense) {
+      return sendMessage(form, `${title.singular} not found!`, 'error');
+    }
+
+    if (expense.attachmentUrls?.length) {
+      await Promise.all(
+        expense.attachmentUrls.map((file) => delFile(file.url)),
+      );
+    }
 
     await db.delete(Expenses).where(eq(Expenses.id, form.data.id));
